main app and login page separated using expo-router
