+++ frontend/src/app/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Redirect } from 'expo-router';
 import { useFonts } from 'expo-font';
 
@@ -7,8 +7,10 @@ export default function Page() {
     MusticaPro: require('../../assets/MusticaPro-SemiBold.otf'),
   });
 
+  const [user, setUser] = useState<boolean>(false);
+
   // See https://docs.expo.dev/develop/user-interface/fonts/#use-a-custom-font if this stops working
 
   if (!fontsLoaded) return;
-  return <Redirect href="/leagues" />;
+  return user ? <Redirect href="/main" /> : <Redirect href="/login" />;
 }
