+++ frontend/src/app/main/bet/_layout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Stack } from 'expo-router';
-import { headerOptions } from '../../styles/styles';
+import { headerOptions } from '../../../styles/styles';
 
 export default function BetLayout() {
   return (
