+++ frontend/src/app/main/bet/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { View, ScrollView } from 'react-native';
-import { globalStyles } from '../../styles/styles';
-import { MatchUp } from '../../components/bet/MatchUp';
-import { Gameweek } from '../../components/bet/Gameweek';
-import { Confirm } from '../../components/bet/Confirm';
+import { globalStyles } from '../../../styles/styles';
+import { MatchUp } from '../../../components/bet/MatchUp';
+import { Gameweek } from '../../../components/bet/Gameweek';
+import { Confirm } from '../../../components/bet/Confirm';
 
 const Bet = () => {
   return (
