+++ frontend/src/app/main/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Redirect } from 'expo-router';
+
+export default function Page() {
+  return <Redirect href="/main/leagues" />;
+}
