--- frontend/src/app/stats/index.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import { View } from 'react-native';
-import { globalStyles } from '../../styles/styles';
-import PremText from '../../components/basic/PremText';
-
-const Stats = () => {
-  return (
-    <View style={globalStyles.container}>
-      <PremText>Some stats</PremText>
-    </View>
-  );
-};
-
-export default Stats;
